[TASK] explizit type cast

+++ src/spi_binding.cc
@@ -192,7 +192,7 @@ void Spi::full_duplex_transfer(
   struct spi_ioc_transfer data = {
 	  (unsigned long)write,
 	  (unsigned long)read,
-	  length,
+	  static_cast<uint32_t>(length),
 	  speed,
 	  delay, // Still unsure ... just expose to options.
 	  bits
